requirements.txt
flask>=2.3.2
google-api-python-client>=2.129.0
google-auth>=2.31.0
google-auth-oauthlib>=1.2.1
python-dateutil>=2.9.0.post0
dateparser>=1.2.0
PyMuPDF>=1.24.2
python-docx>=1.1.2
pytesseract>=0.3.10
Pillow>=10.3.0
beautifulsoup4>=4.12.3 